refactor: move FeelInputParam record to upper level
