+++ connectors/agentic-ai/src/main/java/io/camunda/connector/agenticai/adhoctoolsschema/feel/FeelInputParamExtractor.java
@@ -184,29 +184,4 @@ public class FeelInputParamExtractor {
       super(message);
     }
   }
-
-  public record FeelInputParam(
-      String name,
-      String description,
-      String type,
-      Map<String, Object> schema,
-      Map<String, Object> options) {
-
-    public FeelInputParam(String name) {
-      this(name, null, null, null, null);
-    }
-
-    public FeelInputParam(String name, String description) {
-      this(name, description, null, null, null);
-    }
-
-    public FeelInputParam(String name, String description, String type) {
-      this(name, description, type, null, null);
-    }
-
-    public FeelInputParam(
-        String name, String description, String type, Map<String, Object> schema) {
-      this(name, description, type, schema, null);
-    }
-  }
 }
